60% progress on report generation function

+++ cstaff/ajax-cstaff-delete-menu-item.php
@@ -2,18 +2,21 @@
 session_start();
 include('../conn_db.php');
 if (!empty(($_POST['mitem_id']))) {
-    $target_dir = '/img/menu/';
     $mitem_id = $_POST['mitem_id'];
-    $query = "SELECT mitem_pic FROM mitem WHERE mitem_id = '{$mitem_id}';";
+    $query = "SELECT * FROM mitem WHERE mitem_id = '{$mitem_id}';";
     $result = $mysqli->query($query);
-    $row = mysqli_fetch_array($result);
-    $target_file = $target_dir . $row['mitem_pic'];
-    unlink(SITE_ROOT . $target_file);
-
+    $row = $result->fetch_array();
+    //Set pic name into variable instead of using mysql query;
+    $mitem_pic = $row['mitem_pic'];
     $query = "DELETE FROM mitem WHERE mitem_id = '{$mitem_id}';";
     $result = $mysqli->query($query);
-    $response['server_status'] = 1;
-} else {
-    $response['server_status'] = 0;
+    if ($result) {
+        $target_dir = '/img/menu/';
+        $target_file = $target_dir . $mitem_pic;
+        unlink(SITE_ROOT . $target_file);
+        $response['server_status'] = 1;
+    } else {
+        $response['server_status'] = 0;
+    }
 }
 echo json_encode($response);
